pyproject.toml
[tool.poetry]
name = "cistiers"
version = "0.1.0"
description = ""
authors = ["wertiko1 <144675169+wertiko1@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.6"
pydantic-settings = "^2.6.1"
uvicorn = "^0.32.1"
pydantic = "^2.10.3"
loguru = "^0.7.3"
asyncpg = "^0.30.0"
tortoise-orm = "^0.22.1"
aiohttp = "^3.11.10"
discord-py = "^2.4.0"
oauthlib = "^3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
